+++ src/main/java/com/arcad/employeebook/service/impl/DepartmentService.java
@@ -1,12 +1,9 @@
 package com.arcad.employeebook.service.impl;
 
 import com.arcad.employeebook.elementaryClasses.Department;
-import com.arcad.employeebook.elementaryClasses.Employee;
 import com.arcad.employeebook.exception.DepartmentAlreadyAddedException;
-import com.arcad.employeebook.exception.InputArgsErrorException;
 
 import java.util.List;
-import java.util.Map;
 
 public interface DepartmentService {
 
@@ -18,14 +15,14 @@ public interface DepartmentService {
      * @return Добавляет новый отдел и возвращает краткий отчет по новой записи
      */
 
-    Department addDepartment(String name, Integer salary)
+    Department addDepartment(String name, double salary)
             throws DepartmentAlreadyAddedException;
 
     Department getDepartment(Integer departmentID);
 
-    void editDepartment(Integer inID, String name, Integer salary);
+    void editDepartment(Integer inID, String name, Double salary);
 
     void editDepartment(Integer inID, String name);
 
-    void editDepartment(Integer inID, Integer salary);
+    void editDepartment(Integer inID, double salary);
 }
